plotbase.py: values in units of pixels have to be ints

+++ HarryPlotter/python/plotbase.py
@@ -90,9 +90,9 @@ class PlotBase(processor.Processor):
 
 		# canvas settings
 		self.canvas_options = parser.add_argument_group("Canvas options")
-		self.canvas_options.add_argument("--canvas-width", type=float, default=600.0,
+		self.canvas_options.add_argument("--canvas-width", type=int, default=600,
 		                                 help="Width of the canvas in pixels. [Default: %(default)s]")
-		self.canvas_options.add_argument("--canvas-height", type=float, default=600.0,
+		self.canvas_options.add_argument("--canvas-height", type=int, default=600,
 		                                 help="Height of the canvas in pixels. [Default: %(default)s]")
 
 		#plot formatting
